Refactor vertical navigation to simplify state management and enhance component structure by utilizing useMemo for navigation items and separating mobile and desktop navigation into distinct components.
