+++ lib/utils/navigation.ts
@@ -1,0 +1,87 @@
+import { SECTIONS } from '@/constants/sections';
+
+export interface NavigationItem {
+  id: string;
+  title: string;
+  icon: React.FC<{ className?: string }>;
+}
+
+export const DEFAULT_NAV_ITEMS: NavigationItem[] = [
+  {
+    id: 'section-learning-resources',
+    title: 'Learning Resources',
+    icon: SECTIONS[0].icon,
+  },
+  {
+    id: 'section-developer-tools',
+    title: 'Developer Tools',
+    icon: SECTIONS[1].icon,
+  },
+  {
+    id: 'section-frameworks-&-libraries',
+    title: 'Frameworks & Libraries',
+    icon: SECTIONS[2].icon,
+  },
+  {
+    id: 'section-communities',
+    title: 'Communities',
+    icon: SECTIONS[3].icon,
+  },
+  { id: 'section-blogs', title: 'Blogs', icon: SECTIONS[4].icon },
+];
+
+export function createSearchNavItems(
+  searchQuery: string
+): NavigationItem[] {
+  if (!searchQuery || searchQuery.trim().length === 0) {
+    return DEFAULT_NAV_ITEMS;
+  }
+
+  const sections = document.querySelectorAll('section[id]');
+  const sectionsWithContent = Array.from(sections).filter(
+    (section) => {
+      const gridContainer = section.querySelector('.grid');
+      return gridContainer && gridContainer.children.length > 0;
+    }
+  );
+
+  const sectionIdsWithContent = sectionsWithContent.map(
+    (section) => section.id
+  );
+
+  const searchNavItems = sectionIdsWithContent
+    .filter((id) => id.startsWith('section-'))
+    .map((id) => {
+      const displayName = id
+        .replace('section-', '')
+        .split('-')
+        .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+        .join(' ')
+        .replace(/\b[aA][nN][dD]\b/g, '&');
+
+      const matchingSection = SECTIONS.find(
+        (section) =>
+          section.title.toLowerCase() === displayName.toLowerCase() ||
+          section.title.toLowerCase().replace(' & ', ' and ') ===
+            displayName.toLowerCase()
+      );
+
+      return {
+        id,
+        title: displayName,
+        icon: matchingSection?.icon || SECTIONS[0].icon,
+      };
+    });
+
+  return searchNavItems.length > 0
+    ? searchNavItems
+    : DEFAULT_NAV_ITEMS;
+}
+
+export function scrollToSection(id: string, onComplete?: () => void) {
+  const element = document.getElementById(id);
+  if (element) {
+    element.scrollIntoView({ behavior: 'smooth' });
+    onComplete?.();
+  }
+}
